fix(ci): repair migrated standard test profiles
