+++ scripts/axbuild/src/test/std.rs
@@ -12,7 +12,6 @@ use cargo_metadata::Metadata;
 use crate::support::process::run_cargo_status;
 
 const STD_CRATES_CSV: &str = "scripts/test/std_crates.csv";
-const MIGHT_SLEEP_FILTER: &str = "might_sleep";
 const TASK_INITIALIZATION_FILTER: &str = "task_initialization_precedes_scheduling";
 
 #[derive(Clone, Copy, Debug)]
@@ -28,28 +27,18 @@ const AX_TASK_FEATURE_PROFILES: &[PackageFeatureProfile] = &[
         name: "host-test+multitask+irq-pure",
         features: &["host-test", "multitask", "irq"],
         name_filter: Some("std_tests::"),
-        expected_tests: &["std_tests::axtask_api_constants_hold"],
+        expected_tests: &[
+            "api::std_tests::axtask_api_constants_hold",
+            "api::std_tests::axtask_api_scheduler_name_hold",
+            "api::std_tests::axtask_api_task_registry_functions_exist_hold",
+            "api::std_tests::axtask_api_type_aliases_hold",
+        ],
     },
     PackageFeatureProfile {
         name: "host-test+multitask-task-initialization",
         features: &["host-test", "multitask"],
         name_filter: Some(TASK_INITIALIZATION_FILTER),
         expected_tests: &["api::tests::task_initialization_precedes_scheduling"],
-    },
-    PackageFeatureProfile {
-        name: "host-test+multitask",
-        features: &["host-test", "multitask"],
-        name_filter: Some(MIGHT_SLEEP_FILTER),
-        expected_tests: &["tests::might_sleep_ignores_irq_state_without_irq_feature"],
-    },
-    PackageFeatureProfile {
-        name: "host-test+multitask+preempt+lockdep",
-        features: &["host-test", "multitask", "preempt", "lockdep"],
-        name_filter: Some(MIGHT_SLEEP_FILTER),
-        expected_tests: &[
-            "tests::might_sleep_reports_held_lock_stack",
-            "tests::might_sleep_reports_preempt_disabled_reason",
-        ],
     },
 ];
 
@@ -650,7 +639,7 @@ mod tests {
     }
 
     #[test]
-    fn ax_task_uses_task_initialization_and_might_sleep_feature_profiles() {
+    fn ax_task_uses_pure_and_task_initialization_feature_profiles() {
         let root = PathBuf::from("/tmp/workspace");
         let packages = vec!["ax-task".to_string()];
         let mut runner = FakeCargoRunner::succeeding().with_ax_task_discovery();
@@ -701,42 +690,6 @@ mod tests {
                     "--features",
                     "host-test,multitask",
                     "task_initialization_precedes_scheduling",
-                ],
-                vec![
-                    "test",
-                    "-p",
-                    "ax-task",
-                    "--features",
-                    "host-test,multitask",
-                    "might_sleep",
-                    "--",
-                    "--list",
-                ],
-                vec![
-                    "test",
-                    "-p",
-                    "ax-task",
-                    "--features",
-                    "host-test,multitask",
-                    "might_sleep",
-                ],
-                vec![
-                    "test",
-                    "-p",
-                    "ax-task",
-                    "--features",
-                    "host-test,multitask,preempt,lockdep",
-                    "might_sleep",
-                    "--",
-                    "--list",
-                ],
-                vec![
-                    "test",
-                    "-p",
-                    "ax-task",
-                    "--features",
-                    "host-test,multitask,preempt,lockdep",
-                    "might_sleep",
                 ],
             ]
         );
@@ -855,25 +808,28 @@ mod tests {
     fn profile_discovery_mismatch_fails_without_running_that_profile() {
         let root = PathBuf::from("/tmp/workspace");
         let packages = vec!["ax-task".to_string()];
-        let basic_profile = &AX_TASK_FEATURE_PROFILES[1];
-        let diagnostic_profile = &AX_TASK_FEATURE_PROFILES[2];
+        let pure_profile = &AX_TASK_FEATURE_PROFILES[0];
+        let initialization_profile = &AX_TASK_FEATURE_PROFILES[1];
         let mut runner = FakeCargoRunner::succeeding()
             .with_ax_task_discovery()
-            .with_listing(basic_profile, &["tests::might_sleep_unexpected"])
-            .with_listing(diagnostic_profile, diagnostic_profile.expected_tests);
+            .with_listing(pure_profile, &["api::std_tests::unexpected"])
+            .with_listing(
+                initialization_profile,
+                initialization_profile.expected_tests,
+            );
 
         let failed = run_std_tests(&mut runner, &root, &packages).unwrap();
 
         assert_eq!(failed, vec!["ax-task"]);
         assert!(!runner.invocations.iter().any(|(_, invocation)| {
             invocation
-                == &CargoTestInvocation::for_profile("ax-task", basic_profile, CargoTestAction::Run)
+                == &CargoTestInvocation::for_profile("ax-task", pure_profile, CargoTestAction::Run)
         }));
         assert!(runner.invocations.iter().any(|(_, invocation)| {
             invocation
                 == &CargoTestInvocation::for_profile(
                     "ax-task",
-                    diagnostic_profile,
+                    initialization_profile,
                     CargoTestAction::Run,
                 )
         }));
@@ -910,6 +866,6 @@ mod tests {
         let failed = run_std_tests(&mut runner, &root, &packages).unwrap();
 
         assert_eq!(failed, vec!["ax-task", "ax-api"]);
-        assert_eq!(runner.invocations.len(), 9);
+        assert_eq!(runner.invocations.len(), 5);
     }
 }
